Move ftx and coin gecko clients to api_clients namespace
