--- app/services/sol_prices/ftx_api_client.rb
@@ -1,16 +1,0 @@
-module SolPrices
-  class FtxApiClient
-    def initialize(api_url:)
-      @api_url = URI(api_url)
-    end
-
-    # FTX
-    def send_request(request)
-      Net::HTTP.start(
-        @api_url.hostname,
-        @api_url.port,
-        use_ssl: @api_url.scheme == 'https'
-      ) { |http| http.request(request) }
-    end
-  end
-end
