+++ script/sol_prices/ftx_gather_historical_prices.rb
@@ -11,7 +11,7 @@ include PipelineLogic
 # Create our initial payload with the input values
 initial_payload = {
   exchange: SolPrice.exchanges[:ftx],
-  client: SolPrices::ApiWrappers::Ftx.new,
+  client: FtxClient.new,
   start_time: DateTime.new(2020,7,27), # first day when sol appeared
   end_time: DateTime.current - 1.day,
   resolution: 86400 # interval of returned data
