+++ script/sol_prices/ftx_gather_yesterday_prices.rb
@@ -11,7 +11,7 @@ include PipelineLogic
 # Create our initial payload with the input values
 initial_payload = {
   exchange: SolPrice.exchanges[:ftx],
-  client: SolPrices::ApiWrappers::Ftx.new,
+  client: FtxClient.new,
   start_time: DateTime.current.beginning_of_day - 1.day,
   end_time: DateTime.current.end_of_day - 1.day,
   resolution: 86400 # interval of returned data
