--- test/services/sol_prices/ftx_api_client_test.rb
@@ -1,4 +1,0 @@
-module SolPrices
-  class FtxApiClientTest < ActiveSupport::TestCase
-  end
-end
