+++ test/vcr_cassettes/api_clients/coin_gecko_client_test/test__status.yml
