+++ test/vcr_cassettes/api_clients/ftx_client_test/test__get_market.yml
@@ -1,0 +1,73 @@
+---
+http_interactions:
+- request:
+    method: get
+    uri: https://ftx.com/api/markets/SOL/USD
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+      Accept:
+      - "*/*"
+      User-Agent:
+      - Ruby
+      Host:
+      - ftx.com
+      Ftx-Key:
+      - "<TOKEN>"
+      Ftx-Sign:
+      - '09ab2f6f3bf894f1d92e9b3f718f8a29224cc0d1b611e36cd79b91f33fb3429e'
+      Ftx-Ts:
+      - '1649176098000'
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Date:
+      - Tue, 05 Apr 2022 16:28:19 GMT
+      Content-Type:
+      - application/json
+      Transfer-Encoding:
+      - chunked
+      Connection:
+      - keep-alive
+      Vary:
+      - Accept-Encoding
+      - Accept-Encoding
+      - Origin
+      Cache-Control:
+      - public, max-age=0
+      Etag:
+      - W/"0bfc3c5916ad0ac092f84ec01888eadd"
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains; preload
+      X-Frame-Options:
+      - sameorigin
+      X-Content-Type-Options:
+      - nosniff
+      X-Xss-Protection:
+      - 1; mode=block
+      Content-Security-Policy:
+      - block-all-mixed-content
+      X-Cache-Status:
+      - MISS
+      Cf-Cache-Status:
+      - MISS
+      Expect-Ct:
+      - max-age=604800, report-uri="https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct"
+      Server:
+      - cloudflare
+      Cf-Ray:
+      - 6f73a7fa4e9376c0-LHR
+      Alt-Svc:
+      - h3=":443"; ma=86400, h3-29=":443"; ma=86400
+    body:
+      encoding: ASCII-8BIT
+      string: '{"success":true,"result":{"name":"SOL/USD","enabled":true,"postOnly":false,"priceIncrement":0.0025,"sizeIncrement":0.01,"minProvideSize":0.01,"last":132.9325,"bid":132.89,"ask":132.925,"price":132.925,"type":"spot","baseCurrency":"SOL","quoteCurrency":"USD","underlying":null,"restricted":false,"highLeverageFeeExempt":true,"largeOrderThreshold":5000.0,"change1h":0.022028294633246195,"change24h":0.023306838083873824,"changeBod":0.0040410907168215125,"quoteVolume24h":111773446.3951,"volumeUsd24h":111773446.3951,"priceHigh24h":136.625,"priceLow24h":125.86}}
+
+        '
+  recorded_at: Tue, 05 Apr 2022 16:28:19 GMT
+recorded_with: VCR 6.0.0
