--- test/vcr_cassettes/services/api_wrappers/ftx/test__get_market.yml
@@ -1,68 +1,0 @@
----
-http_interactions:
-- request:
-    method: get
-    uri: https://ftx.com/api/markets/SOL/USD
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-      Host:
-      - ftx.com
-      Ftx-Key:
-      - AUk9zEtHT7Jan9XINgEBpMOzp-XP7F28qExTYYj1
-      Ftx-Sign:
-      - 2a291810474344a41a03d6ddd402a601f1ac210454b31e5b14ed53dc4ff11dba
-      Ftx-Ts:
-      - '1630926446000'
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Mon, 06 Sep 2021 11:07:26 GMT
-      Content-Type:
-      - application/json
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Vary:
-      - Accept-Encoding
-      - Accept-Encoding
-      - Origin
-      Cache-Control:
-      - public, max-age=0
-      Etag:
-      - W/"a7173fe8cd7298b50861c16a464db601"
-      X-Cache-Status:
-      - MISS
-      Cf-Cache-Status:
-      - MISS
-      Expect-Ct:
-      - max-age=604800, report-uri="https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct"
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains; preload
-      X-Content-Type-Options:
-      - nosniff
-      Server:
-      - cloudflare
-      Cf-Ray:
-      - 68a73bd0fd1df2a8-WAW
-      Alt-Svc:
-      - h3-27=":443"; ma=86400, h3-28=":443"; ma=86400, h3-29=":443"; ma=86400, h3=":443";
-        ma=86400
-    body:
-      encoding: ASCII-8BIT
-      string: '{"success":true,"result":{"name":"SOL/USD","enabled":true,"postOnly":false,"priceIncrement":0.0025,"sizeIncrement":0.01,"minProvideSize":0.01,"last":143.9275,"bid":143.935,"ask":144.055,"price":143.935,"type":"spot","baseCurrency":"SOL","quoteCurrency":"USD","underlying":null,"restricted":false,"highLeverageFeeExempt":true,"change1h":-0.010569007888088814,"change24h":0.026603901430048857,"changeBod":0.013430497614898524,"quoteVolume24h":131157310.1566,"volumeUsd24h":131157310.1566}}
-
-        '
-  recorded_at: Mon, 06 Sep 2021 11:07:26 GMT
-recorded_with: VCR 6.0.0
